Use a base_channel_cut to spare a lot of redundant lines. The drawback might be the intialization of a few quantities irrelevant for some topologies. The "accept" method cannot be generalized as the pattern interface is very specific (e.g. get_electrons_internal_probability). However, with this structure, it should be possible to add other quantities on which to cut (in the channel_2eNg_cut for intance, one could add a cut on the internal probability between an electron and a gamma different than the existing one between two electrons)
